+++ src/main/java/com/whatachad/app/model/domain/DateTime.java
@@ -1,0 +1,17 @@
+package com.whatachad.app.model.domain;
+
+import jakarta.persistence.Embeddable;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Embeddable
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class DateTime {
+    private Integer date;
+    private Integer hour;
+    private Integer minute;
+}
