--- src/main/java/com/whatachad/app/model/domain/Workcheck.java
@@ -1,5 +1,0 @@
-package com.whatachad.app.model.domain;
-
-public enum Workcheck {
-    COMPLETE, YET
-}
