+++ src/main/java/com/whatachad/app/type/Workcheck.java
@@ -1,0 +1,5 @@
+package com.whatachad.app.type;
+
+public enum Workcheck {
+    COMPLETE, NOT_COMPLETE
+}
